分类完成

+++ src/main/java/io/ukoko/bhwms/service/impl/CategoryServiceImpl.java
@@ -1,6 +1,7 @@
 package io.ukoko.bhwms.service.impl;
 
 import io.ukoko.bhwms.entity.Category;
+import io.ukoko.bhwms.exceptions.BhWmsException;
 import io.ukoko.bhwms.mapper.CategoryMapper;
 import io.ukoko.bhwms.service.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,7 +32,13 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void deleteCategory(Integer categoryId) {
-        categoryMapper.deleteCategory(categoryId);
+    public void deleteCategory(Integer categoryId) throws BhWmsException{
+        //判断当前分类下是否存在叶子节点,如果存在不能删除
+        List<Category> categories = categoryMapper.getCategoryListByParentId(categoryId);
+        if(categories!=null && categories.size()>0){
+            throw new BhWmsException(3,"存在叶子节点,不能删除");
+        }else{
+            categoryMapper.deleteCategory(categoryId);
+        }
     }
 }
